Prefs widget now "works"
Added `prefs.js` and the necessary `buildPrefsWidget()` function so that
the extension actually has a preferences pane that loads "correctly"
when opened in `gnome-tweak-tool`. This seems to _work_, but it does not
look very good. The extension itself is still not loading, but at least
now the actual restults are engaged.

Also note, that if the settings schema is changed at all
`glib-compile-schemas` must be run in order to update
`gschemas.compiled`.